Update outcome status label and move `Status` enum
Changes the label name to reduce the possibility of a clash with other labels.

Moves `Status` class under `OutcomeRecorder` as it is only used there.
